WIP: Design and Implementation of the ui module
